feat: todo service 구조 변경
