+++ src/main/java/NotModified/TodoList/dto/todo/TodoGroupedResponseDto.java
@@ -17,13 +17,15 @@ public class TodoGroupedResponseDto {
     private String categoryColor;
     private Boolean categoryIsActive;
     private List<TodoResponseDto> todos = new ArrayList<>();
+    private String lesson;
 
-    public TodoGroupedResponseDto(Long categoryId, String categoryName, String categoryColor, Boolean categoryIsActive) {
+    public TodoGroupedResponseDto(Long categoryId, String categoryName, String categoryColor, Boolean categoryIsActive, String lesson) {
         this.categoryId = categoryId;
         this.categoryName = categoryName;
         this.categoryColor = categoryColor;
         this.categoryIsActive = categoryIsActive;
         this.todos = new ArrayList<>(); // 초기화
+        this.lesson = lesson;
     }
 
     public void addTodo(Todo todo) {
