+++ src/main/java/NotModified/TodoList/service/TodoService.java
@@ -1,9 +1,11 @@
 package NotModified.TodoList.service;
 
 import NotModified.TodoList.domain.Category;
+import NotModified.TodoList.domain.Lesson;
 import NotModified.TodoList.domain.Todo;
 import NotModified.TodoList.dto.todo.*;
 import NotModified.TodoList.repository.CategoryRepository;
+import NotModified.TodoList.repository.LessonRepository;
 import NotModified.TodoList.repository.TodoRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -18,6 +20,7 @@ public class TodoService {
     private final TodoRepository todoRepository;
     private final CategoryRepository categoryRepository;
 
+    // todo 등록
     public Long saveTodo(TodoCreateRequestDto todoDto) {
         // 존재하지 않거나, 비활성화된 카테고리에 대해서는 등록할 수 없음
         Category category = categoryRepository.findById(todoDto.getCategoryId())
@@ -34,32 +37,8 @@ public class TodoService {
         todoRepository.save(todo);
         return todo.getId();
     }
-
-    public List<TodoGroupedResponseDto> findDateList(TodoDateRequestDto todoDto) {
-        List<Todo> todos = todoRepository.findByDate(todoDto.getTodoDate(), todoDto.getUserId());
-
-        Map<String, TodoGroupedResponseDto> groupedMap = new LinkedHashMap<>();
-
-        // 특정 날짜의 todo 목록을 추출
-        for (Todo todo : todos) {
-            // todo 가 속한 category 분류의 id 값 가져옴
-            Long categoryId = todo.getCategory().getId();
-
-            // 그룹 키 = categoryId
-            String key = String.valueOf(categoryId);
-
-            // key 값이 존재하지 않으면 새로운 그룹화 배열을 생성
-            groupedMap.computeIfAbsent(key, k -> new TodoGroupedResponseDto(categoryId, todo.getCategory().getName(), todo.getCategory().getColor(), todo.getCategory().getIsActive()));
-            // 해당하는 categoryName 그룹에 Todo 객체를 추가
-            groupedMap.get(key).addTodo(todo);
-        }
-        return new ArrayList<>(groupedMap.values());
-    }
-
-    public List<Todo> findTodos() {
-        return todoRepository.findAll();
-    }
-
+    
+    // todo 수정
     public void updateTodo(Long id, TodoUpdateRequestDto dto) {
         Todo todo = todoRepository.findById(id)
                 .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 todo 입니다."));
@@ -67,6 +46,7 @@ public class TodoService {
         if(dto.getIsDone() != null) todo.setIsDone(dto.getIsDone());
     }
 
+    // todo 삭제
     public void deleteTodo(Long id) {
         Todo todo = todoRepository.findById(id)
                 .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 todo 입니다."));
